+++ src/adaptor/source/string.rs
@@ -20,9 +20,12 @@ impl<'a, T: Parser<&'a str>> StringSource<'a, T> {
 
 impl<'a, T: Parser<&'a str>> Source for StringSource<'a, T> {
     fn parse(&self) -> Result<Value, RealmError> {
-        Value::try_serialize(
-            &T::parse(self.buffer)
-                .map_err(|e| RealmError::ParseError(e.to_string()))?,
-        )
+        Value::try_serialize(&T::parse(self.buffer).map_err(|e| {
+            RealmError::new_parse_error(
+                self.buffer.to_string(),
+                "string".to_string(),
+                e.to_string(),
+            )
+        })?)
     }
 }